% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdDbClass.R
\docType{methods}
\name{getAlleles,IpdDb-method}
\alias{getAlleles,IpdDb-method}
\alias{getAlleles}
\title{Get alleles}
\usage{
getAlleles(x, locus)
}
\arguments{
\item{x}{The database connection; an \code{\link{IpdDb}} object.}

\item{locus}{A single locus as a string.}
}
\value{
A character vector with all alleles of the give locus.
}
\description{
Get all alleles of a given locus.
}
\examples{
## Load the database 
hla <- loadHlaData()
## Get the loci
loci <- getLoci(hla)
## Get alleles of a locus
alleles <- getAlleles(hla, loci[1])

}
