% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdDbClass.R
\docType{methods}
\name{getLoci,IpdDb-method}
\alias{getLoci,IpdDb-method}
\alias{getLoci}
\title{Get loci}
\usage{
getLoci(x)
}
\arguments{
\item{x}{The database connection; an \code{\link{IpdDb}} object.}
}
\value{
A vector of available loci in the database.
}
\description{
Get all available loci of the KIR or HLA database
}
\examples{
## Load the database 
hla <- loadHlaData()
## Get the loci
loci <- getLoci(hla)

}
