% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IpdDb.R
\docType{package}
\name{ipdDbPackage}
\alias{ipdDbPackage}
\alias{ipdDbPackage-package}
\title{Get allele information from IPD}
\description{
This package holds the IPD IMGT/HLA and IPD KIR database.
All alleles are accessible using the select, columns, keys and keytypes
methods of the AnnoatationDbi package of bioconductor.
}
\details{
Included data are:
  
Allele names

p-groups

g-groups

cwd_status

completeness status

gene structure

reference sequences

closest full-length allele
}
