% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDb.R
\name{loadHlaData}
\alias{loadHlaData}
\title{Load the IPD IMGT/HLA database}
\usage{
loadHlaData(version = "Latest")
}
\arguments{
\item{version}{Either a valid version of the IPD IMGT/HLA database or 
"Latest" to fetch the latest version}
}
\value{
an \code{\link{IpdDb}} object containing the database.
}
\description{
Load the IPD IMGT/HLA database
}
\examples{
## Load the HLA database
hla <- loadHlaData()
}
