% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDb.R
\name{loadKirData}
\alias{loadKirData}
\title{Load the IPD KIR database}
\usage{
loadKirData()
}
\value{
an \code{\link{IpdDb}} object containing the database.
}
\description{
Load the IPD KIR database
}
\examples{
hla <- loadKirData()
}
